モデルの定義を分割
