+++ lib/ui/components/dialog/event/add_event_dialog.dart
@@ -3,7 +3,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:mr_collection/constants/base_url.dart';
-import 'package:mr_collection/data/model/freezed/lineGroup.dart';
+import 'package:mr_collection/data/model/freezed/line_group.dart';
 import 'package:mr_collection/provider/user_provider.dart';
 import 'package:mr_collection/data/repository/event_repository.dart';
 import 'package:mr_collection/ui/screen/transfer/choice_event_screen.dart';
@@ -85,7 +85,7 @@ class AddEventDialogState extends ConsumerState<AddEventDialog> {
       if (_isTransferMode) {
         await ref.read(userProvider.notifier).createEventAndTransferMembers(
             _selectedEvent!.eventId, eventName, userId);
-      } else if(lineGroup != null){
+      } else if (lineGroup != null) {
         await ref.read(userProvider.notifier).createEventAndGetMembersFromLine(
             lineGroup!.groupId, eventName, lineGroup!.members, userId);
       } else {
@@ -113,27 +113,26 @@ class AddEventDialogState extends ConsumerState<AddEventDialog> {
 
   Future<void> _selectLineGroup() async {
     final userId = ref.read(userProvider)?.userId;
-    if( userId == null ) return;
+    if (userId == null) return;
 
-    final lineGroups = await ref.read(userProvider.notifier).getLineGroups(userId);
+    final lineGroups =
+        await ref.read(userProvider.notifier).getLineGroups(userId);
     debugPrint('LINE取得APIを実行しました。');
     debugPrint('取得したLINEグループ : $lineGroups');
 
-    if(lineGroups.isEmpty){
+    if (lineGroups.isEmpty) {
       Navigator.of(context).push(
         MaterialPageRoute(
-          // 集金くんアカウントを含むLINEグループがないとき
-            builder: (_) => const InviteOfficialAccountToLineGroupScreen()
-        ),
+            // 集金くんアカウントを含むLINEグループがないとき
+            builder: (_) => const InviteOfficialAccountToLineGroupScreen()),
       );
     }
     final pickedLineGroup = await Navigator.of(context).push<LineGroup>(
-        MaterialPageRoute(
+      MaterialPageRoute(
           // 集金くんアカウントを含むLINEグループがあるとき
-          builder: (_) => SelectLineGroupScreen(lineGroups : lineGroups)
-        ),
-      );
-    if(pickedLineGroup != null){
+          builder: (_) => SelectLineGroupScreen(lineGroups: lineGroups)),
+    );
+    if (pickedLineGroup != null) {
       setState(() {
         lineGroup = pickedLineGroup;
       });
@@ -286,41 +285,43 @@ class AddEventDialogState extends ConsumerState<AddEventDialog> {
                                   color: Colors.black),
                         ),
                         trailing: lineGroup != null
-                        ? SizedBox(
-                            width: 112,
-                            height: 28,
-                            child: ElevatedButton(
-                              onPressed: null,
-                              style: ElevatedButton.styleFrom(
-                                backgroundColor: const Color(0xFF06C755),
-                                elevation: 2,
-                                shape: const StadiumBorder(),
-                                padding: const EdgeInsets.symmetric(
-                                  horizontal: 4, vertical: 0),
-                              ),
-                              child: Text(
-                                lineGroup!.groupName,
-                                maxLines: 1,
-                                style: Theme.of(context)
-                                    .textTheme
-                                    .bodyMedium
-                                    ?.copyWith(
-                                  fontSize: 10,
-                                  fontWeight: FontWeight.w500,
-                                  color: _isTransferMode
-                                      ? Colors.white
-                                      : Colors.black,
+                            ? SizedBox(
+                                width: 112,
+                                height: 28,
+                                child: ElevatedButton(
+                                  onPressed: null,
+                                  style: ElevatedButton.styleFrom(
+                                    backgroundColor: const Color(0xFF06C755),
+                                    elevation: 2,
+                                    shape: const StadiumBorder(),
+                                    padding: const EdgeInsets.symmetric(
+                                        horizontal: 4, vertical: 0),
+                                  ),
+                                  child: Text(
+                                    lineGroup!.groupName,
+                                    maxLines: 1,
+                                    style: Theme.of(context)
+                                        .textTheme
+                                        .bodyMedium
+                                        ?.copyWith(
+                                          fontSize: 10,
+                                          fontWeight: FontWeight.w500,
+                                          color: _isTransferMode
+                                              ? Colors.white
+                                              : Colors.black,
+                                        ),
+                                  ),
+                                ))
+                            : IconButton(
+                                icon: SvgPicture.asset(
+                                  'assets/icons/line.svg',
+                                  width: 32,
+                                  height: 32,
                                 ),
+                                onPressed: _isButtonEnabled
+                                    ? () => _selectLineGroup()
+                                    : null,
                               ),
-                            )
-                        ): IconButton(
-                          icon: SvgPicture.asset(
-                            'assets/icons/line.svg',
-                            width: 32,
-                            height: 32,
-                          ),
-                          onPressed: _isButtonEnabled ? () => _selectLineGroup() : null,
-                        ),
                       ),
                     ),
                   ],
