+++ lib/ui/screen/line_add_member/select_line_group_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:mr_collection/data/model/freezed/lineGroup.dart';
+import 'package:mr_collection/data/model/freezed/line_group.dart';
 import 'package:mr_collection/provider/user_provider.dart';
 import 'package:mr_collection/data/model/freezed/event.dart';
 import 'package:mr_collection/ui/components/dialog/invite_official_acount_to_line_group_dialog.dart';
@@ -10,18 +10,19 @@ import 'package:flutter_gen/gen_l10n/s.dart';
 
 class SelectLineGroupScreen extends StatelessWidget {
   final List<LineGroup> lineGroups;
-  const SelectLineGroupScreen({Key? key, required this.lineGroups}) : super(key: key);
+  const SelectLineGroupScreen({Key? key, required this.lineGroups})
+      : super(key: key);
 
-  Future<void> _checkSelectedLineGroup(BuildContext context, LineGroup lineGroup) async {
+  Future<void> _checkSelectedLineGroup(
+      BuildContext context, LineGroup lineGroup) async {
     Navigator.of(context).push<Event>(
-    MaterialPageRoute(
-    builder: (_) => CheckSelectedLineGroupScreen(lineGroup: lineGroup)),
+      MaterialPageRoute(
+          builder: (_) => CheckSelectedLineGroupScreen(lineGroup: lineGroup)),
     );
   }
 
   @override
   Widget build(BuildContext context) {
-
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: AppBar(
@@ -51,16 +52,18 @@ class SelectLineGroupScreen extends StatelessWidget {
         actions: [
           GestureDetector(
             onTap: () {
-                showDialog(
-                  context: context,
-                  builder: (context) => const InviteOfficialAccountToLineGroupDialog(),
-                );
-              },
+              showDialog(
+                context: context,
+                builder: (context) =>
+                    const InviteOfficialAccountToLineGroupDialog(),
+              );
+            },
             child: Row(
               mainAxisAlignment: MainAxisAlignment.start,
               children: [
                 Text(
-                  S.of(context)?.notDisplayedQuestion ?? "LINE group not displayed?",
+                  S.of(context)?.notDisplayedQuestion ??
+                      "LINE group not displayed?",
                   style: Theme.of(context).textTheme.bodyMedium?.copyWith(
                       color: const Color(0xFF5C5C5C),
                       fontSize: 10,
@@ -71,7 +74,8 @@ class SelectLineGroupScreen extends StatelessWidget {
                   'assets/icons/question_circle.svg',
                   width: 24,
                   height: 24,
-                  colorFilter: const ColorFilter.mode(Color(0xFF8E8E93), BlendMode.srcIn),
+                  colorFilter: const ColorFilter.mode(
+                      Color(0xFF8E8E93), BlendMode.srcIn),
                 ),
                 const SizedBox(width: 24),
               ],
@@ -83,60 +87,63 @@ class SelectLineGroupScreen extends StatelessWidget {
         padding: const EdgeInsets.symmetric(horizontal: 24),
         child: Column(crossAxisAlignment: CrossAxisAlignment.center, children: [
           Text(
-            S.of(context)?.selectLineGroupTitle ?? "Add members from LINE group",
+            S.of(context)?.selectLineGroupTitle ??
+                "Add members from LINE group",
             style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-              fontSize: 24,
-              fontWeight: FontWeight.w700,
-              color: const Color(0xFF06C755),
-              height: 1.1,
-            ),
+                  fontSize: 24,
+                  fontWeight: FontWeight.w700,
+                  color: const Color(0xFF06C755),
+                  height: 1.1,
+                ),
             textAlign: TextAlign.center,
           ),
           const SizedBox(height: 32),
           Text(
-            S.of(context)?.selectLineGroupDesc ?? "Select the LINE group where you want to add members.",
+            S.of(context)?.selectLineGroupDesc ??
+                "Select the LINE group where you want to add members.",
             style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-              fontSize: 16,
-              fontWeight: FontWeight.w700,
-              color: Colors.black,
-            ),
+                  fontSize: 16,
+                  fontWeight: FontWeight.w700,
+                  color: Colors.black,
+                ),
             textAlign: TextAlign.center,
           ),
           const SizedBox(height: 32),
           Expanded(
               child: ListView.builder(
-                itemCount: lineGroups.length,
-                itemBuilder: (context, index) {
-                  final lineGroup = lineGroups[index];
-                  return Column(
-                    children: [
-                      SizedBox(
-                        height: 44,
-                        child: ListTile(
-                            title: Text(
-                              lineGroup.groupName,
-                              style:
+            itemCount: lineGroups.length,
+            itemBuilder: (context, index) {
+              final lineGroup = lineGroups[index];
+              return Column(
+                children: [
+                  SizedBox(
+                    height: 44,
+                    child: ListTile(
+                        title: Text(
+                          lineGroup.groupName,
+                          style:
                               Theme.of(context).textTheme.bodyMedium?.copyWith(
-                                fontSize: 16,
-                                fontWeight: FontWeight.w500,
-                              ),
-                            ),
-                            trailing: SizedBox(
-                              width: 16,
-                              height: 16,
-                              child: SvgPicture.asset('assets/icons/ic_next.svg'),
-                            ),
-                            onTap: () => _checkSelectedLineGroup(context, lineGroup)),
-                      ),
-                      const Divider(
-                        color: Color(0xFFE8E8E8),
-                        thickness: 1,
-                        height: 1,
-                      ),
-                    ],
-                  );
-                },
-              )),
+                                    fontSize: 16,
+                                    fontWeight: FontWeight.w500,
+                                  ),
+                        ),
+                        trailing: SizedBox(
+                          width: 16,
+                          height: 16,
+                          child: SvgPicture.asset('assets/icons/ic_next.svg'),
+                        ),
+                        onTap: () =>
+                            _checkSelectedLineGroup(context, lineGroup)),
+                  ),
+                  const Divider(
+                    color: Color(0xFFE8E8E8),
+                    thickness: 1,
+                    height: 1,
+                  ),
+                ],
+              );
+            },
+          )),
         ]),
       ),
     );
